- moves authentication providers to their target environment
