+++ src/Kiota.Builder/SearchProviders/GitHub/Autentication/AccessCodeResponse.cs
@@ -3,7 +3,7 @@ using System.Text.Json.Serialization;
 
 namespace Kiota.Builder.SearchProviders.GitHub.Authentication;
 
-internal class AccessCodeResponse {
+public class AccessCodeResponse {
 	[JsonPropertyName("access_token")]
 	public string AccessToken { get; set; }
 	[JsonPropertyName("token_type")]
